+++ src/Ensembles/Ensembles.jl
@@ -1,141 +1,130 @@
 """
     Ensembles
 
-This module provides the main function [`run_ensemble`](@ref Ensembles.run_ensemble).
+This module provides the main function [`run_dynamics`](@ref run_dynamics).
 This serves to run multiple trajectories for a given simulation type, sampling
 from an initial distribution.
 """
 module Ensembles
 
-using SciMLBase: SciMLBase, EnsembleProblem, solve
-using RecursiveArrayTools: ArrayPartition
-using TypedTables: TypedTables
+using SciMLBase: SciMLBase
+using Dictionaries: Dictionary
+using UnitfulAtomic: austrip
 
 using NQCBase: NQCBase
 using NQCDynamics:
     AbstractSimulation,
-    Simulation,
-    RingPolymerSimulation,
-    DynamicsUtils,
-    DynamicsMethods,
-    DynamicsOutputs
-using RingPolymerArrays: RingPolymerArray
+    DynamicsMethods
 
-export run_ensemble
+export run_dynamics
 
 include("selections.jl")
-include("outputs.jl")
+
 include("reductions.jl")
+export SumReduction
+export MeanReduction
+export AppendReduction
+export FileReduction
 
 """
-    run_ensemble(sim::AbstractSimulation, tspan, distribution;
+    EnsembleSaver{F<:Tuple}
+
+Store a tuple of functions with the signature `f(sol)` where `sol` is a DiffEq solution object.
+`EnsembleSaver` will evaluate each of these functions and return the result in a `Dictionary`.
+"""
+struct EnsembleSaver{F<:Tuple}
+    functions::F
+end
+
+function (output::EnsembleSaver)(sol, i)
+    out = Dictionary{Symbol,Any}([:Time], [sol.t])
+    return (evaluate_output_functions!(out, sol, i, output.functions...), false)
+end
+
+function evaluate_output_functions!(out::Dictionary, sol, i, f::F, fs...) where {F}
+    if f isa Function
+        name = nameof(f)
+    else
+        name = nameof(typeof(f))
+    end
+
+    insert!(out, name, f(sol, i))
+    return evaluate_output_functions!(out, sol, i, fs...)
+end
+evaluate_output_functions!(out::Dictionary, sol, i) = out
+
+"""
+    run_dynamics(sim::AbstractSimulation, tspan, distribution;
+        output,
         selection::Union{Nothing,AbstractVector}=nothing,
-        output=(sol,i)->(sol,false),
-        reduction=:append
-        reduction::Symbol=:append,
-        ensemble_algorithm=SciMLBase.EnsembleThreads(),
+        reduction=AppendReduction(),
+        ensemble_algorithm=SciMLBase.EnsembleSerial(),
         algorithm=DynamicsMethods.select_algorithm(sim),
-        saveat=[],
         trajectories=1,
         kwargs...
         )
 
-Run multiple trajectories for timespan `tspan` sampling from `distribution`.
-The DifferentialEquations ensemble interface is used which allows us to specify
-functions to modify the output and how it is reduced across trajectories.
+Run trajectories for timespan `tspan` sampling from `distribution`.
 
 # Keywords
 
+* `output` either a single function or a Tuple of functions with the signature `f(sol, i)` that takes the DifferentialEquations solution and returns the desired output quantity.
 * `selection` should be an `AbstractVector` containing the indices to sample from the `distribution`. By default, `nothing` leads to random sampling.
-* `output` can be a function that transforms the DiffEq solution to an output, or a tuple of output quantities as for `run_trajectory`.
-* `reduction` defines how the data is reduced across trajectories. Options are `:append`, `:mean` or `:sum`.
+* `reduction` defines how the data is reduced across trajectories. Options are `AppendReduction()`, `MeanReduction()`, `SumReduction` and `FileReduction(filename)`.
 * `ensemble_algorithm` is the algorithm from DifferentialEquations which determines which form of parallelism is used.
 * `algorithm` is the algorithm used to integrate the equations of motion.
-* `saveat` mirrors the DiffEq keyword and labels the time points to save the output.
 * `trajectories` is the number of trajectories to perform.
-
-This function wraps the EnsembleProblem from DifferentialEquations and passes the `kwargs`
-to the `solve` function.
+* `kwargs...` any additional keywords are passed to DifferentialEquations `solve``.
 """
-function run_ensemble(
+function run_dynamics(
     sim::AbstractSimulation, tspan, distribution;
+    output,
     selection::Union{Nothing,AbstractVector}=nothing,
-    output=SciMLBase.DEFAULT_OUTPUT_FUNC,
-    reduction::Symbol=:append,
-    ensemble_algorithm=SciMLBase.EnsembleThreads(),
+    reduction=AppendReduction(),
+    ensemble_algorithm=SciMLBase.EnsembleSerial(),
     algorithm=DynamicsMethods.select_algorithm(sim),
-    saveat=[],
     trajectories=1,
-    u_init=nothing,
     kwargs...
 )
 
+    if !(output isa Tuple)
+        output = (output,)
+    end
+
     kwargs = NQCBase.austrip_kwargs(;kwargs...)
     trajectories = convert(Int, trajectories)
-    saveat = austrip.(saveat)
     tspan = austrip.(tspan)
-    (output isa Symbol) && (output = (output,))
-    reduction = Reduction(reduction)
 
-    ensemble_problem = EnsembleProblem(sim, tspan, distribution, selection, output, reduction, u_init, saveat, trajectories, kwargs)
+    prob_func = Selection(distribution, selection, trajectories)
 
-    @info "Performing $trajectories trajectories."
-    stats = @timed solve(ensemble_problem, algorithm, ensemble_algorithm; saveat, trajectories, kwargs...)
+    u0 = sample_distribution(sim, distribution)
+    problem = DynamicsMethods.create_problem(u0, tspan, sim)
+
+    output_func = EnsembleSaver(output)
+
+    ensemble_problem = SciMLBase.EnsembleProblem(problem; prob_func, output_func, reduction)
+
+    if trajectories == 1
+        @info "Performing 1 trajectory."
+    else
+        @info "Performing $trajectories trajectories."
+    end
+
+    stats = @timed SciMLBase.solve(ensemble_problem, algorithm, ensemble_algorithm; trajectories, kwargs...)
     @info "Finished after $(stats.time) seconds."
 
-    return extract_output(stats, output, reduction, trajectories)
-end
-
-function SciMLBase.EnsembleProblem(sim::AbstractSimulation, tspan, distribution, selection, output, reduction, u_init, saveat, trajectories, kwargs)
-
-    prob_func = Selection(distribution, selection)
-
-    problem = create_problem(sim, distribution, tspan)
-    output_func = Output(output, sim)
-    if output_func isa DynamicsOutputs.EnsembleSaver
-        u_init = get_u_init(reduction, saveat, kwargs, tspan, problem.u0, output_func)
-    end
-
-    return EnsembleProblem(problem; prob_func, output_func, reduction, u_init)
-end
-
-Output(output::Tuple, sim) = DynamicsOutputs.EnsembleSaver(output, sim)
-Output(output, _) = output
-
-function create_problem(sim::AbstractSimulation, distribution, tspan)
-    u0 = sample_distribution(sim, distribution)
-    return DynamicsMethods.create_problem(u0, tspan, sim)
-end
-
-function extract_output(stats, output, reduction, trajectories)
-    sol = stats.value
-
-    if output isa Tuple
-        process_output(reduction, sol, output, trajectories)
+    if trajectories == 1
+        return stats.value.u[1]
     else
-        return sol.u
+        return stats.value.u
     end
 end
 
-function process_output(::AppendReduction, sol, output, trajectories)
-    return [generate_table(output, s) for s in sol.u]
-end
-
-function process_output(::SumReduction, sol, output, trajectories)
-    for i in eachindex(sol.u)
-        sol.u[i][1] /= trajectories
-    end
-    return generate_table(output, sol.u)
-end
-
-function process_output(::MeanReduction, sol, output, trajectories)
-    return generate_table(output, sol.u)
-end
-
-function generate_table(output, u)
-    output = (:t, output...)
-    out = [(;zip(output, quantity)...) for quantity in u]
-    return TypedTables.Table(out)
-end
+run_ensemble(args...; kwargs...) = throw(error("""
+`run_ensemble` has been replaced, use `run_dynamics` instead.
+`run_dynamics` unifies single trajectory and ensemble simulations into one function.
+Refer to the `run_dynamics` docstring for more information.
+"""
+))
 
 end # module